+++ lib/generators/neighbor/vector_generator.rb
@@ -2,12 +2,12 @@ require "rails/generators/active_record"
 
 module Neighbor
   module Generators
-    class InstallGenerator < Rails::Generators::Base
+    class VectorGenerator < Rails::Generators::Base
       include ActiveRecord::Generators::Migration
       source_root File.join(__dir__, "templates")
 
       def copy_migration
-        migration_template "migration.rb", "db/migrate/install_neighbor.rb", migration_version: migration_version
+        migration_template "vector.rb", "db/migrate/install_neighbor_vector.rb", migration_version: migration_version
       end
 
       def migration_version