+++ lib/neighbor.rb
@@ -7,10 +7,14 @@ require "neighbor/version"
 module Neighbor
   class Error < StandardError; end
 
-  module RegisterCubeType
+  module RegisterTypes
     def initialize_type_map(m = type_map)
       super
       m.register_type "cube", ActiveRecord::ConnectionAdapters::PostgreSQL::OID::SpecializedString.new(:cube)
+      m.register_type "vector" do |_, _, sql_type|
+        limit = extract_limit(sql_type)
+        ActiveRecord::ConnectionAdapters::PostgreSQL::OID::SpecializedString.new(:vector, limit: limit)
+      end
     end
   end
 end
@@ -25,16 +29,20 @@ ActiveSupport.on_load(:active_record) do
 
   # ensure schema can be dumped
   ActiveRecord::ConnectionAdapters::PostgreSQLAdapter::NATIVE_DATABASE_TYPES[:cube] = {name: "cube"}
+  ActiveRecord::ConnectionAdapters::PostgreSQLAdapter::NATIVE_DATABASE_TYPES[:vector] = {name: "vector"}
 
   # ensure schema can be loaded
   if ActiveRecord::VERSION::MAJOR >= 6
-    ActiveRecord::ConnectionAdapters::TableDefinition.send(:define_column_methods, :cube)
+    ActiveRecord::ConnectionAdapters::TableDefinition.send(:define_column_methods, :cube, :vector)
   else
     ActiveRecord::ConnectionAdapters::TableDefinition.define_method :cube do |*args, **options|
       args.each { |name| column(name, :cube, options) }
     end
+    ActiveRecord::ConnectionAdapters::TableDefinition.define_method :vector do |*args, **options|
+      args.each { |name| column(name, :vector, options) }
+    end
   end
 
   # prevent unknown OID warning
-  ActiveRecord::ConnectionAdapters::PostgreSQLAdapter.prepend(Neighbor::RegisterCubeType)
+  ActiveRecord::ConnectionAdapters::PostgreSQLAdapter.prepend(Neighbor::RegisterTypes)
 end