+++ CHANGELOG.md
@@ -1,8 +1,15 @@
-## 0.1.3 (unreleased)
+## 0.2.0 (unreleased)
 
+- Added support for pgvector
+- Added `normalize` option
+- Made `dimensions` optional
 - Raise an error if `nearest_neighbors` already defined
 - Raise an error for non-finite values
 - Fixed NaN with zero vectors and cosine distance
+
+Breaking changes
+
+- The `distance` option has been moved from `has_neighbors` to `nearest_neighbors`, and there is no longer a default
 
 ## 0.1.2 (2021-02-21)